+++ srcs/exec/call_builtin_command.c
@@ -2,7 +2,7 @@
 #include "ms_builtin.h"
 #include "ft_string.h"
 
-int	call_builtin_command(const char *const *argv, t_params *params)
+uint8_t	call_builtin_command(const char *const *argv, t_params *params)
 {
 	if (ft_streq(argv[0], CMD_ECHO))
 		return (ft_echo(argv));
